[test] multi-backend ci
