+++ .circleci/config.yml
@@ -36,9 +36,16 @@ jobs:
       - run:
           command: python setup.py install
       - run:
-          name: Run tests
+          name: Run TF tests
           # This assumes pytest is installed via the install-package step above
-          command: pytest
+          command: TL_BACKEND=tensorflow pytest
+      - run:
+          name: Run TH tests
+          command: TL_BACKEND=torch pytest
+      - run:
+          name: Run PD tests
+          command: TL_BACKEND=paddle pytest
+
 
 # Invoke jobs via workflows
 # See: https://circleci.com/docs/2.0/configuration-reference/#workflows
