+++ gammagl/datasets/hgb.py
@@ -108,7 +108,7 @@ class HGBDataset(InMemoryDataset):
         url = self.url.format(self.names[self.name])
         path = download_url(url, self.raw_dir, self.names[self.name]+'.zip')
         extract_zip(path, self.raw_dir)
-        shutil.rmtree(osp.join(self.raw_dir,"__MACOSX"))
+        shutil.rmtree(osp.join(self.raw_dir, "__MACOSX"))
         for filename in self.raw_file_names:
             filePath = osp.join(self.raw_dir,self.names[self.name],filename)
             shutil.move(filePath, self.raw_dir)
@@ -165,7 +165,7 @@ class HGBDataset(InMemoryDataset):
                 x_dict[n_type].append([float(v) for v in x[3].split(',')])
         for n_type in n_types.values():
             if len(x_dict[n_type]) == 0:
-                data[n_type].x = tlx.ops.eye(num_nodes_dict[n_type], num_nodes_dict[n_type])
+                data[n_type].x = tlx.ops.eye(num_nodes_dict[n_type])
                 data[n_type].num_nodes = num_nodes_dict[n_type]
             else:
                 data[n_type].x = tlx.ops.convert_to_tensor(x_dict[n_type])
