+++ tests/layers/conv/test_agnn_conv.py
@@ -8,6 +8,7 @@
 import tensorlayerx as tlx
 from gammagl.layers.conv import AGNNConv
 
+
 def test_agnn_conv():
     x = tlx.random_normal(shape=(4, 64))
     edge_index = tlx.convert_to_tensor([[0, 0, 0, 1, 2, 3], [1, 2, 3, 0, 0, 0]])
@@ -15,5 +16,6 @@ def test_agnn_conv():
     conv = AGNNConv(in_channels=64)
     out = conv(x, edge_index=edge_index, num_nodes=4)
     assert out.shape == (4, 64)
+
 
 test_agnn_conv()
