+++ tests/utils/test_coalesce.py
@@ -1,6 +1,3 @@
-import os
-os.environ['TL_BACKEND'] = 'tensorflow'
-
 import tensorlayerx as tlx
 from gammagl.utils import coalesce
 
@@ -22,5 +19,3 @@ def test_coalesce():
     assert tlx.ops.convert_to_numpy(edge_index).tolist() == [[0, 1], [1, 0]]
     assert tlx.ops.convert_to_numpy(attr).tolist() == [[-4.0, 4.0], [2.0, 5.0]]
 
-
-# test_coalesce()
