+++ tests/utils/test_degree.py
@@ -1,6 +1,3 @@
-import os
-os.environ['TL_BACKEND'] = 'tensorflow'
-
 import tensorlayerx as tlx
 from gammagl.utils import degree
 
@@ -9,6 +6,5 @@ def test_degree():
 	row = tlx.convert_to_tensor([0, 1, 0, 2, 0])
 	deg = degree(row, dtype=tlx.int64)
 	assert deg.dtype == tlx.int64
-	assert tlx.ops.convert_to_numpy(deg).tolist() == [3, 1, 1]
+	assert tlx.convert_to_numpy(deg).tolist() == [3, 1, 1]
 
-# test_degree()
