+++ tests/utils/test_sort_edge_index.py
@@ -1,6 +1,3 @@
-import os
-os.environ['TL_BACKEND'] = 'tensorflow'
-
 import tensorlayerx as tlx
 
 from gammagl.utils import sort_edge_index
@@ -16,5 +13,3 @@ def test_sort_edge_index():
     out = sort_edge_index(edge_index, edge_attr)
     assert tlx.ops.convert_to_numpy(out[0]).tolist() == [[0, 1, 1, 2], [1, 0, 2, 1]]
     assert tlx.ops.convert_to_numpy(out[1]).tolist() == [[4], [3], [2], [1]]
-
-# test_sort_edge_index()
